% Generated by roxygen2 (4.0.2): do not edit by hand
\name{to_coordinates}
\alias{to_coordinates}
\title{Turn a matrix of values into a data frame with x and y
coordinate columns and a value column.}
\usage{
to_coordinates(mat)
}
\description{
Turn a matrix of values into a data frame with x and y
coordinate columns and a value column.
}

